modeling RestaurantController & RestaurantRegisterService and refactoring KitchenController & KitchenRegisterService

+++ src/main/java/com/evereats/fooder/infrastructure/repository/RestaurantRepositoryImpl.java
@@ -2,10 +2,12 @@ package com.evereats.fooder.infrastructure.repository;
 
 import com.evereats.fooder.domain.model.Restaurant;
 import com.evereats.fooder.domain.repository.RestaurantRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Component
@@ -25,13 +27,20 @@ public class RestaurantRepositoryImpl implements RestaurantRepository {
     }
 
     @Override
-    public Restaurant add(Restaurant restaurant) {
+    @Transactional
+    public Restaurant save(Restaurant restaurant) {
         return entityManager.merge(restaurant);
     }
 
     @Override
-    public void remove(Restaurant restaurant) {
-        restaurant = findById(restaurant.getId());
+    @Transactional
+    public void delete(Long id) {
+        Restaurant restaurant = findById(id);
+
+        if (restaurant == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
+
         entityManager.remove(restaurant);
     }
 }
